Move token classes to Tokens folder
